+++ question/bank/tagquestion/amd/src/selectors.js
@@ -16,13 +16,12 @@
 /**
  * The purpose of this module is to centralize selectors related to question.
  *
- * @module     core_question/question_selectors
+ * @module     qbank_tagquestion/question_selectors
+ * @package    qbank_tagquestion
  * @copyright  2018 Simey Lameze <lameze@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 define([], function() {
-    // eslint-disable-next-line no-console
-    console.log('warn: The core_question/selectors has been deprecated. Please use qbank_tagquestion/selectors instead.');
     return {
         actions: {
             save: '[data-action="save"]',
